Fix bug generic adapter

+++ adapter/generic_adapter.go
@@ -4,15 +4,20 @@ import (
 	"context"
 	"database/sql"
 	"database/sql/driver"
+	"encoding/json"
+	"errors"
 	"fmt"
-	q "github.com/core-go/sql"
 	"reflect"
+
+	q "github.com/core-go/sql"
 )
 
 type GenericAdapter[T any, K any] struct {
 	*Adapter[*T]
 	Map    map[string]int
 	Fields string
+	Keys   []string
+	IdMap  bool
 }
 
 func NewGenericAdapter[T any, K any](db *sql.DB, tableName string, opts ...func(int) string) (*GenericAdapter[T, K], error) {
@@ -29,17 +34,31 @@ func NewSqlGenericAdapterWithVersionAndArray[T any, K any](db *sql.DB, tableName
 	if err != nil {
 		return nil, err
 	}
+
 	var t T
 	modelType := reflect.TypeOf(t)
-	if modelType.Kind() == reflect.Ptr {
-		modelType = modelType.Elem()
+	if modelType.Kind() != reflect.Struct {
+		return nil, errors.New("T must be a struct")
 	}
+
+	_, primaryKeys := q.FindPrimaryKeys(modelType)
+	var k K
+	kType := reflect.TypeOf(k)
+	idMap := false
+	if len(primaryKeys) > 1 {
+		if kType.Kind() == reflect.Map {
+			idMap = true
+		} else if kType.Kind() != reflect.Struct {
+			return nil, errors.New("For composite keys, K must be a struct or a map")
+		}
+	}
+
 	fieldsIndex, err := q.GetColumnIndexes(modelType)
 	if err != nil {
 		return nil, err
 	}
 	fields := q.BuildFieldsBySchema(adapter.Schema)
-	return &GenericAdapter[T, K]{adapter, fieldsIndex, fields}, nil
+	return &GenericAdapter[T, K]{adapter, fieldsIndex, fields, primaryKeys, idMap}, nil
 }
 func (a *GenericAdapter[T, K]) All(ctx context.Context) ([]T, error) {
 	var objs []T
@@ -48,10 +67,31 @@ func (a *GenericAdapter[T, K]) All(ctx context.Context) ([]T, error) {
 	err := q.Query(ctx, tx, a.Map, &objs, query)
 	return objs, err
 }
+func toMap(obj interface{}) (map[string]interface{}, error) {
+	b, err := json.Marshal(obj)
+	if err != nil {
+		return nil, err
+	}
+	im := make(map[string]interface{})
+	er2 := json.Unmarshal(b, &im)
+	return im, er2
+}
+func (a *GenericAdapter[T, K]) getId(k K) (interface{}, error) {
+	if len(a.Keys) >= 2 && !a.IdMap {
+		ri, err := toMap(k)
+		return ri, err
+	} else {
+		return k, nil
+	}
+}
 func (a *GenericAdapter[T, K]) Load(ctx context.Context, id K) (*T, error) {
+	ip, er0 := a.getId(id)
+	if er0 != nil {
+		return nil, er0
+	}
 	var objs []T
 	query := fmt.Sprintf("select %s from %s ", a.Fields, a.Table)
-	query1, args := q.BuildFindByIdWithDB(a.DB, query, id, a.JsonColumnMap, a.Schema.SKeys, a.BuildParam)
+	query1, args := q.BuildFindByIdWithDB(a.DB, query, ip, a.JsonColumnMap, a.Schema.SKeys, a.BuildParam)
 	tx := q.GetExec(ctx, a.DB, a.TxKey)
 	err := q.Query(ctx, tx, a.Map, &objs, query1, args...)
 	if err != nil {
@@ -63,8 +103,12 @@ func (a *GenericAdapter[T, K]) Load(ctx context.Context, id K) (*T, error) {
 	return nil, nil
 }
 func (a *GenericAdapter[T, K]) Exist(ctx context.Context, id K) (bool, error) {
+	ip, er0 := a.getId(id)
+	if er0 != nil {
+		return false, er0
+	}
 	query := fmt.Sprintf("select %s from %s ", a.Schema.SColumns[0], a.Table)
-	query1, args := q.BuildFindByIdWithDB(a.DB, query, id, a.JsonColumnMap, a.Schema.SKeys, a.BuildParam)
+	query1, args := q.BuildFindByIdWithDB(a.DB, query, ip, a.JsonColumnMap, a.Schema.SKeys, a.BuildParam)
 	tx := q.GetExec(ctx, a.DB, a.TxKey)
 	rows, err := tx.QueryContext(ctx, query1, args...)
 	if err != nil {
@@ -77,8 +121,12 @@ func (a *GenericAdapter[T, K]) Exist(ctx context.Context, id K) (bool, error) {
 	return false, nil
 }
 func (a *GenericAdapter[T, K]) Delete(ctx context.Context, id K) (int64, error) {
+	ip, er0 := a.getId(id)
+	if er0 != nil {
+		return -1, er0
+	}
 	query := fmt.Sprintf("delete from %s ", a.Table)
-	query1, args := q.BuildFindByIdWithDB(a.DB, query, id, a.JsonColumnMap, a.Schema.SKeys, a.BuildParam)
+	query1, args := q.BuildFindByIdWithDB(a.DB, query, ip, a.JsonColumnMap, a.Schema.SKeys, a.BuildParam)
 	tx := q.GetExec(ctx, a.DB, a.TxKey)
 	res, err := tx.ExecContext(ctx, query1, args...)
 	if err != nil {
